Moved fetching logic
